backend/requirements.txt
# Hospital Management System - Backend Dependencies

# Core Python packages
setuptools>=68.0.0
wheel>=0.41.0

# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Authorization
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# API Documentation
drf-spectacular==0.26.5

# Environment Variables
python-decouple==3.8

# Date & Time
python-dateutil==2.8.2

# Validation
django-phonenumber-field==7.2.0
phonenumbers==8.13.25

# File Handling
Pillow==10.1.0

# Production Server
gunicorn==21.2.0

# Development Tools
django-extensions==3.2.3

# Testing
factory-boy==3.3.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Security
django-ratelimit==4.1.0

# Utilities
django-filter==23.4
django-crispy-forms==2.1
crispy-bootstrap5==0.7
